[CP-2425] Extracted history properties (HistoryLocationState) in redirect-to-discovery.action
