+++ libs/core/app-initialization/hooks/use-initializing-app-effects.ts
@@ -5,16 +5,13 @@
 
 import { useEffect } from "react"
 import { useDispatch, useSelector } from "react-redux"
-import { useHistory } from "react-router-dom"
 import { startInitializingApp } from "Core/app-initialization/actions/start-initializing-app"
 import { Dispatch } from "Core/__deprecated__/renderer/store"
 import { setAppInitializationStatus } from "Core/app-initialization/actions/base.action"
 import { AppInitializationStatus } from "Core/app-initialization/reducers/app-initialization.interface"
 import { isAppUpdateProcessPassed } from "Core/app-initialization/selectors/is-app-update-process-passed.selector"
-import { redirectToDiscovery } from "Core/app-initialization/actions/redirect-to-discovery.action"
 
 export const useInitializingAppEffects = () => {
-  const history = useHistory()
   const dispatch = useDispatch<Dispatch>()
   const appUpdateProcessPassed = useSelector(isAppUpdateProcessPassed)
 
@@ -25,7 +22,6 @@ export const useInitializingAppEffects = () => {
   useEffect(() => {
     if (appUpdateProcessPassed) {
       dispatch(setAppInitializationStatus(AppInitializationStatus.Initialized))
-      dispatch(redirectToDiscovery(history))
     }
-  }, [history, dispatch, appUpdateProcessPassed])
+  }, [dispatch, appUpdateProcessPassed])
 }
