+++ libs/core/core/components/apps/base-app/base-app.component.tsx
@@ -16,6 +16,7 @@ import { useWatchUnlockStatus } from "Core/core/hooks/use-watch-unlock-status-ef
 import { useDeviceLockedEffect } from "Core/core/hooks/use-device-locked-effect"
 import { useDeviceDetachedEffect } from "Core/core/hooks/use-device-detached-effect"
 import { useDeviceConnectFailedEffect } from "Core/core/hooks/use-device-connect-failed-effect"
+import { useDiscoveryRedirectEffect } from "Core/core/hooks/use-discovery-redirect-effect"
 import { useRouterListener } from "Core/core/hooks"
 import { useAPISerialPortListeners, useOutbox } from "generic-view/store"
 
@@ -29,6 +30,7 @@ const BaseApp: FunctionComponent = () => {
   useDeviceLockedEffect()
   useWatchOutboxEntriesEffect()
   useWatchUnlockStatus()
+  useDiscoveryRedirectEffect()
   // API
   useAPISerialPortListeners()
   useOutbox()
